utils/requirements.txt
absl-py==0.7.0
alembic==1.1.0
antigravity==0.1
anytree==2.6.0
apache-airflow==1.10.4
apispec==2.0.2
appdirs==1.4.3
appnope==0.1.0
asn1crypto==0.24.0
astor==0.7.1
atomicwrites==1.3.0
attrs==18.2.0
Automat==0.7.0
Babel==2.7.0
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==3.1.0
boto==2.49.0
boto3==1.11.9
botocore==1.14.9
cached-property==1.5.1
cachetools==3.1.1
certifi==2019.11.28
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
colorama==0.4.1
colorlog==4.0.2
configparser==3.5.3
constantly==15.1.0
croniter==0.3.30
cryptography==2.5
cssselect==1.0.3
cycler==0.10.0
decorator==4.3.2
defusedxml==0.6.0
dewiki==0.1.0
dill==0.2.9
Django==2.2.6
docutils==0.15.2
dumb-init==1.2.2
entrypoints==0.3
Flask==1.1.1
Flask-Admin==1.5.3
Flask-AppBuilder==1.13.1
Flask-Babel==0.12.2
Flask-Caching==1.3.3
Flask-JWT-Extended==3.21.0
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.0
flask-swagger==0.2.13
Flask-WTF==0.14.2
fsspec==0.6.2
funcsigs==1.0.0
funcy==1.14
future==0.16.0
Fuzzy==1.2.2
fuzzywuzzy==0.17.0
gast==0.2.2
gensim==3.8.1
google-api-core==1.14.3
google-api-python-client==1.7.11
google-auth==1.4.2
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-storage==1.22.0
google-colab==1.0.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
graphviz==0.10.1
grpcio==1.16.1
gunicorn==19.9.0
gym==0.15.3
h5py==2.9.0
holidays==0.9.9
html5lib==0.9999999
httplib2==0.15.0
hyperlink==18.0.0
idna==2.8
imageio==2.6.1
importlib==1.0.4
incremental==17.5.0
ipykernel==5.1.3
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
iso8601==0.1.12
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.14.1
json-merge-patch==0.2
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lazy-object-proxy==1.4.2
lightgbm==2.2.3
lockfile==0.12.2
lxml==4.3.1
Mako==1.1.0
Markdown==2.6.11
MarkupSafe==1.1.0
marshmallow==2.19.5
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.17.0
matplotlib==3.0.2
mistune==0.8.4
mkl-fft==1.0.10
mkl-random==1.0.2
mock==2.0.0
more-itertools==5.0.0
nbconvert==5.3.1
nbformat==4.4.0
nbmerge==0.0.4
networkx==2.4
nltk==3.4.5
notebook==5.2.2
numexpr==2.7.1
numpy==1.15.4
oauth2client==4.1.3
opencv-python==4.1.1.26
ordereddict==1.1
pandas==0.24.1
pandocfilters==1.4.2
parsel==1.5.1
parso==0.3.2
patsy==0.5.1
pbr==5.1.3
pendulum==1.4.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
plotly==3.7.0
pluggy==0.8.1
portpicker==1.2.0
prison==0.1.0
prometheus-client==0.5.0
prompt-toolkit==1.0.18
protobuf==3.6.1
psutil==5.6.3
psycopg2==2.8.3
ptyprocess==0.6.0
py==1.7.0
py4j==0.10.7
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyDispatcher==2.0.5
PyDrive==1.3.1
pyglet==1.3.2
Pygments==2.3.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pyLDAvis==2.1.2
pyOpenSSL==19.0.0
pyparsing==2.3.1
pyrsistent==0.15.4
pyspark==2.4.0
pyspellchecker==0.5.3
pytest==4.2.1
pytest-runner==4.2
python-daemon==2.1.2
python-dateutil==2.7.5
python-editor==1.0.4
python3-openid==3.1.0
pytz==2018.9
pytzdata==2019.2
PyWavelets==1.1.1
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
queuelib==1.5.0
regex==2019.6.8
requests==2.21.0
retrying==1.3.3
rsa==4.0
s3fs==0.4.0
s3transfer==0.3.2
scikit-image==0.16.2
scikit-learn==0.20.3
scikit-surprise==1.1.0
scipy==1.2.0
Scrapy==1.5.1
seaborn==0.9.0
Send2Trash==1.5.0
service-identity==18.1.0
setproctitle==1.1.10
simplegeneric==0.8.1
six==1.12.0
sklearn==0.0
smart-open==1.9.0
soupsieve==1.9.4
SQLAlchemy==1.3.8
sqlparse==0.3.0
statsmodels==0.9.0
stevedore==1.31.0
tabulate==0.8.3
tenacity==4.12.0
tensorboard==1.13.1
tensorflow==1.5.0
tensorflow-estimator==1.13.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
text-unidecode==1.2
thrift==0.11.0
torch==1.3.0
torchvision==0.4.1
tornado==4.5.3
tqdm==4.31.1
traitlets==4.3.2
Twisted==17.5.0
tzlocal==1.5.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.24.1
virtualenv==16.7.5
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
w3lib==1.20.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.4.2
words==0.0.1
WTForms==2.2.1
xgboost==0.82
xlrd==1.2.0
zope.deprecation==4.4.0
zope.interface==4.6.0
